New direction after this
Currently the program tracks cell IDs but the data analysis is only
being performed on the final population of cells and their label
distributions. It is not enough to only do it at the end. A history of
each cell’s lineage is required, i.e a binary tree structure tracking
the lineage of each cell from the initial population is needed. The
data analysis class should then be responsible for doing the final
output and further analysis of these structures.